Refactor: ororganisms/auth 디렉토리 삭제
