--- src/components/organisms/auth/index.js
@@ -1,3 +1,0 @@
-export { default as LoginCheck } from './LoginCheck';
-export { default as Signin } from './Signin';
-export { default as Signup } from './Signup';
